Refactor hoist-react admin: + Make config a top-level tab + New 'User Data' tab to hold prefs and json blobs
