+++ admin/AppModel.js
@@ -8,10 +8,11 @@ import {TabContainerModel} from '@xh/hoist/cmp/tab';
 import {HoistAppModel, managed, XH} from '@xh/hoist/core';
 import {Icon} from '@xh/hoist/icon';
 import {activityTab} from './tabs/activity/ActivityTab';
+import {configTab} from './tabs/config/ConfigTab';
 import {generalTab} from './tabs/general/GeneralTab';
 import {loggingTab} from './tabs/logging/LoggingTab';
 import {monitorTab} from './tabs/monitor/MonitorTab';
-import {preferencesTab} from './tabs/preferences/PreferencesTab';
+import {userDataTab} from './tabs/userData/UserDataTab';
 
 @HoistAppModel
 export class AppModel {
@@ -50,12 +51,15 @@ export class AppModel {
                 path: '/general',
                 children: [
                     {name: 'about', path: '/about'},
-                    {name: 'config', path: '/config'},
                     {name: 'services', path: '/services'},
                     {name: 'ehCache', path: '/ehCache'},
                     {name: 'users', path: '/users'},
                     {name: 'webSockets', path: '/webSockets'}
                 ]
+            },
+            {
+                name: 'config',
+                path: 'config'
             },
             {
                 name: 'logging',
@@ -83,8 +87,8 @@ export class AppModel {
                 ]
             },
             {
-                name: 'preferences',
-                path: '/preferences',
+                name: 'userData',
+                path: '/userData',
                 children: [
                     {name: 'prefs', path: '/prefs'},
                     {name: 'userPrefs', path: '/userPrefs'},
@@ -97,10 +101,11 @@ export class AppModel {
     createTabs() {
         return [
             {id: 'general', icon: Icon.info(), content: generalTab},
+            {id: 'config', icon: Icon.settings(), content: configTab},
             {id: 'activity', icon: Icon.analytics(), content: activityTab},
             {id: 'logging', icon: Icon.fileText(), content: loggingTab},
             {id: 'monitor', icon: Icon.shieldCheck(), content: monitorTab, omit: !XH.getConf('xhEnableMonitoring', true)},
-            {id: 'preferences', icon: Icon.bookmark(), content: preferencesTab}
+            {id: 'userData', icon: Icon.users(), content: userDataTab}
         ];
     }
 }
