+++ admin/tabs/userData/JsonBlobPanel.js
