+++ admin/tabs/userData/UserDataTab.js
@@ -11,10 +11,10 @@ import {preferencePanel} from './PreferencePanel';
 import {userPreferencePanel} from './UserPreferencePanel';
 import {jsonBlobPanel} from './JsonBlobPanel';
 
-export const preferencesTab = hoistCmp.factory(
+export const userDataTab = hoistCmp.factory(
     () => tabContainer({
         model: {
-            route: 'default.preferences',
+            route: 'default.userData',
             switcherPosition: 'left',
             tabs: [
                 {id: 'prefs', icon: Icon.bookmark(), content: preferencePanel},
